requirements.txt
aiohttp>=3.11.11
google-generativeai>=0.8.3
google-api-python-client>=2.157.0
google-auth>=2.37.0
typer>=0.15.1
rich>=13.9.4
gitpython>=3.1.44
esprima>=4.0.1
javalang>=0.13.0
python-dotenv>=1.0.1
requests>=2.32.3
tqdm>=4.67.1
pydantic>=2.10.4
ratelimit>=2.2.1
cryptography>=44.0.0

# Development dependencies
pytest>=8.3.4
black>=24.10.0
isort>=5.13.2
mypy>=1.8.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
